fix: `apply_multiple_validators` function was not working after run `ruff-check` hook
Ruff changed the comparison `if result != True` to `if not result`.

In most of the cases this makes sense and it's more 'pythonic', but in our scenario, the `result` variable will not be ever an boolean, sometimes it's an string containing the error message, and while this variable its not `True` we keep printing the error message to user, forcing him to fix the error. When the error is fixed, `result` variable stores an boolean `True` So that the why we need to check if its `not True`.