fix: enforces `commit type`  prompt be required
The only workaround for this, since the Questionary 'select' prompt type havent
an `validate` argument, was pass and kwarg called `validator`. That kwarg is
internally passed to `prompt_toolkit.PromptSession` that has an `validator` argument
and this argument expects an class with an `validate` method.

So for keep the consistency on the code based I converted all validators to classes with
an `validate` method.